+++ src/main/java/Main/classes/PropertyForRent.java
@@ -5,15 +5,17 @@ import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Entity
 @Data
 @DiscriminatorValue("3")
 public class PropertyForRent extends Property{
     @Column(name = "pre_cost")
-    private Integer preCost;
+    private BigInteger preCost;
 
     @Column(name = "monthly_rent")
-    private Integer monthlyRent;
+    private BigInteger monthlyRent;
 
     public PropertyForRent() {
     }
