+++ src/main/java/Main/classes/PropertyForSale.java
@@ -6,13 +6,15 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 @Entity
 @DiscriminatorValue("1")
 public class PropertyForSale extends Property {
 
     @Column(name = "sell_cost")
-    private Integer sellCost;
+    private BigInteger sellCost;
 
     public PropertyForSale() {
     }
